@@ -1,46 +1,54 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup Label="Package">
-    <PackageId>DashAccountingSystemV2.Tests</PackageId>
-    <Version>2.0.0</Version>
-    <Authors>Geoffrey Roberts</Authors>
-    <Company>Dash Software Solutions, Inc.</Company>
-    <Copyright>Copyright (C) 2022 Dash Software Solutions, Inc.  All rights reserved.</Copyright>
+	<PackageId>DashAccountingSystemV2.Tests</PackageId>
+	<Version>2.1.0</Version>
+	<Authors>Geoffrey Roberts</Authors>
+	<Company>Dash Software Solutions, Inc.</Company>
+	<Copyright>Copyright (C) 2022 - 2024 Dash Software Solutions, Inc.  All rights reserved.</Copyright>
   </PropertyGroup>
-  
+
   <PropertyGroup Label="Build">
-    <TargetFramework>net5.0</TargetFramework>
-    <UserSecretsId>338122a3-0812-4421-ba78-e20c312529da</UserSecretsId>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
+    <IsTestProject>true</IsTestProject>
+    <UserSecretsId>215dfb3f-4958-402e-9016-4d21e4dd8f83</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup Label="Package References">
-    <PackageReference Include="Dapper" Version="2.0.90" />
-    <PackageReference Include="Microsoft.AspNetCore.ApiAuthorization.IdentityServer" Version="5.0.7" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="5.0.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
-    <PackageReference Include="Moq" Version="4.15.2" />
-    <PackageReference Include="Npgsql" Version="5.0.7" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.7" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    <PackageReference Include="coverlet.collector" Version="6.0.2">
       <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="coverlet.collector" Version="1.3.0">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.7" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="8.0.0" />
+    <PackageReference Include="Microsoft.Extensions.TimeProvider.Testing" Version="8.7.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+    <PackageReference Include="Moq" Version="4.20.70" />
+    <PackageReference Include="Npgsql" Version="8.0.3" />
+    <PackageReference Include="xunit" Version="2.9.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2">
       <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
   </ItemGroup>
 
   <ItemGroup Label="Project References">
-    <ProjectReference Include="..\..\src\DashAccountingSystemV2.csproj" />
+    <ProjectReference Include="..\..\src\BackEnd\DashAccountingSystemV2.BackEnd.csproj" />
+  </ItemGroup>
+
+  <ItemGroup Label="Implicit Usings">
+    <Using Include="Xunit" />
   </ItemGroup>
 
   <ItemGroup Label="Resources">
     <None Update="appsettings.UnitTests.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
 
