@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Xunit;
-using DashAccountingSystemV2.Extensions;
+using DashAccountingSystemV2.BackEnd.Extensions;
 
 namespace DashAccountingSystemV2.Tests.Extensions
 {
@@ -35,7 +31,7 @@ namespace DashAccountingSystemV2.Tests.Extensions
         [InlineData("XY", "****XY", 3)]
         [InlineData("SuperSecretPassword", "****word")]
         [InlineData("SuperSecretPassword", "****Password", 8)]
-        public void MaskSecretValue_Ok(string originalInput, string expectedResult, int? numberOfCharactersToReveal = null)
+        public void MaskSecretValue_Ok(string? originalInput, string? expectedResult, int? numberOfCharactersToReveal = null)
         {
             var actualResult = numberOfCharactersToReveal.HasValue ?
                 originalInput.MaskSecretValue(numberOfCharactersToReveal.Value) :
@@ -47,13 +43,16 @@ namespace DashAccountingSystemV2.Tests.Extensions
         [Fact]
         public void ParseCommaSeparatedIntegers_Ok()
         {
-            string input = null;
+            string? input = null;
             var results = input.ParseCommaSeparatedIntegers();
             Assert.Null(results);
 
             input = "1,3,5,7,9,11, 13, 15,  19";
             results = input.ParseCommaSeparatedIntegers();
+            
+            Assert.NotNull(results);
             Assert.Equal(9, results.Count());
+            
             var expectedValues = new uint[] { 1, 3, 5, 7, 9, 11, 13, 15, 19 };
 
             foreach (var expected in expectedValues)
@@ -64,12 +63,12 @@ namespace DashAccountingSystemV2.Tests.Extensions
         public void ParseListOfCommaSeparatedIntegers_Ok()
         {
             // Case 1: Input is NULL
-            List<string> input = null;
+            List<string>? input = null;
             var results = input.ParseCommaSeparatedIntegers();
             Assert.Null(results);
 
             // Case 2: Input is Empty
-            input = new List<string>();
+            input = [];
             results = input.ParseCommaSeparatedIntegers();
             Assert.Null(results);
 
@@ -92,7 +91,7 @@ namespace DashAccountingSystemV2.Tests.Extensions
         [InlineData("Hablas Español?", "Hablas Espanol?")]
         [InlineData("Können Sie mir behilflich sein?", "Koennen Sie mir behilflich sein?")]
         [InlineData("Ich weiß nicht", "Ich weiss nicht")]
-        public void RemoveDiactricits_Ok(string originalInput, string expectedResult)
+        public void RemoveDiacritics_Ok(string originalInput, string expectedResult)
         {
             var actualResult = originalInput.RemoveDiacritics();
             Assert.Equal(expectedResult, actualResult);
@@ -126,19 +125,19 @@ namespace DashAccountingSystemV2.Tests.Extensions
         [InlineData("Dash Software Solutions, Inc.", "Dash_Software_Solutions_Inc", false, "_")]
         [InlineData("Können Sie mir behilflich sein?", "koennen-sie-mir-behilflich-sein")]
         [InlineData("Ich weiß nicht", "ich-weiss-nicht")]
-        public void Slugify_Ok(string originalInput, string expectedResult, bool? forceLowerCase = true, string delimiter = null)
+        public void Slugify_Ok(string originalInput, string expectedResult, bool? forceLowerCase = true, string? delimiter = null)
         {
             var forceLowercaseSpecified = forceLowerCase.HasValue;
-            var delimiterSpecified = delimiter.HasValue();
+            var delimiterSpecified = !string.IsNullOrEmpty(delimiter);
 
-            string actualResult = null;
+            string? actualResult = null;
 
             if (forceLowercaseSpecified && delimiterSpecified)
-                actualResult = originalInput.Slugify(forceLowerCase.Value, delimiter);
+                actualResult = originalInput.Slugify(forceLowerCase!.Value, delimiter!);
             else if (forceLowercaseSpecified)
-                actualResult = originalInput.Slugify(forceLowerCase.Value);
+                actualResult = originalInput.Slugify(forceLowerCase!.Value);
             else if (delimiterSpecified)
-                actualResult = originalInput.Slugify(delimiter: delimiter);
+                actualResult = originalInput.Slugify(delimiter: delimiter!);
 
             Assert.Equal(expectedResult, actualResult);
         }
