@@ -1,6 +1,5 @@
-﻿using Xunit;
-using DashAccountingSystemV2.Extensions;
-using DashAccountingSystemV2.Models;
+﻿using DashAccountingSystemV2.BackEnd.Extensions;
+using DashAccountingSystemV2.BackEnd.Models;
 
 namespace DashAccountingSystemV2.Tests.Extensions
 {
