@@ -1,12 +1,8 @@
-﻿using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Dapper;
 using Npgsql;
-using Xunit;
-using DashAccountingSystemV2.Models;
-using DashAccountingSystemV2.Repositories;
+using DashAccountingSystemV2.BackEnd.Models;
+using DashAccountingSystemV2.BackEnd.Repositories;
 
 namespace DashAccountingSystemV2.Tests.Repositories
 {
@@ -30,7 +26,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
         private TimeActivity _timeActivity2;
         private TimeActivity _timeActivity3;
 
-        private static readonly DateTime _timeActivitiesDate = new DateTime(2022, 2, 17, 0, 0, 0, DateTimeKind.Utc);
+        private static readonly DateTime _timeActivitiesDate = new DateTime(2022, 2, 17, 0, 0, 0, DateTimeKind.Unspecified);
 
         [Fact]
         [Trait("Category", "Requires Database")]
@@ -45,7 +41,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
                     DepositAccountId = _cashOperatingAccount.Id,
                     RevenueAccountId = _revenueAccount.Id,
                     PaymentMethodId = _paymentMethodCheck.Id,
-                    Date = new DateTime(2022, 2, 27, 0, 0, 0, DateTimeKind.Utc),
+                    Date = new DateTime(2022, 2, 27, 0, 0, 0, DateTimeKind.Unspecified),
                     Amount = _invoice.Total,
                     AssetTypeId = _assetTypeUSD.Id,
                     CheckNumber = 12345,
@@ -271,9 +267,9 @@ namespace DashAccountingSystemV2.Tests.Repositories
             using (var connection = new NpgsqlConnection(connString))
             {
                 // Make a Tenant
-                _tenantId = connection.QueryFirstOrDefault<Guid>($@"
-                    INSERT INTO ""Tenant"" ( ""Name"" )
-                    VALUES ( 'Unit Testing {Guid.NewGuid()}, Inc.' )
+                _tenantId = await connection.QueryFirstOrDefaultAsync<Guid>($@"
+                    INSERT INTO ""Tenant"" ( ""Name"", ""DefaultAssetTypeId"" )
+                    VALUES ( 'Unit Testing {Guid.NewGuid()}, Inc.', 1 )
                     RETURNING ""Id"";");
 
                 // Get a User to use
@@ -502,7 +498,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
             _paymentMethodCheck = paymentMethods.FirstOrDefault(pm => pm.Name == "Check");
 
             // Initialize Payment Journal Entry
-            var entryDate = new DateTime(2022, 2, 27, 0, 0, 0, DateTimeKind.Utc);
+            var entryDate = new DateTime(2022, 2, 27, 0, 0, 0, DateTimeKind.Unspecified);
 
             var journalEntry = new JournalEntry(
                 _tenantId,
