@@ -1,13 +1,9 @@
-﻿using System;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Dapper;
 using Npgsql;
-using Xunit;
-using DashAccountingSystemV2.Extensions;
-using DashAccountingSystemV2.Models;
-using DashAccountingSystemV2.Repositories;
+using DashAccountingSystemV2.BackEnd.Extensions;
+using DashAccountingSystemV2.BackEnd.Models;
+using DashAccountingSystemV2.BackEnd.Repositories;
 
 namespace DashAccountingSystemV2.Tests.Repositories
 {
@@ -32,7 +28,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
                     CustomerId = _customer.EntityId,
                     EmployeeId = _employee.EntityId,
                     ProductId = _product.Id,
-                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Utc),
+                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Unspecified),
                     StartTime = new TimeSpan(9, 0, 0),
                     EndTime = new TimeSpan(16, 0, 0),
                     Break = new TimeSpan(2, 0, 0),
@@ -84,7 +80,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
                     CustomerId = _customer.EntityId,
                     EmployeeId = _employee.EntityId,
                     ProductId = _product.Id,
-                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Utc),
+                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Unspecified),
                     StartTime = new TimeSpan(9, 0, 0),
                     EndTime = new TimeSpan(16, 0, 0),
                     Break = new TimeSpan(2, 0, 0),
@@ -132,7 +128,7 @@ namespace DashAccountingSystemV2.Tests.Repositories
                     CustomerId = _customer.EntityId,
                     EmployeeId = _employee.EntityId,
                     ProductId = _product.Id,
-                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Utc),
+                    Date = new DateTime(2021, 7, 14, 0, 0, 0, DateTimeKind.Unspecified),
                     StartTime = new TimeSpan(9, 0, 0),
                     EndTime = new TimeSpan(16, 0, 0),
                     Break = new TimeSpan(2, 0, 0),
@@ -209,9 +205,9 @@ namespace DashAccountingSystemV2.Tests.Repositories
             using (var connection = new NpgsqlConnection(connString))
             {
                 // Make a Tenant
-                _tenantId = connection.QueryFirstOrDefault<Guid>($@"
-                    INSERT INTO ""Tenant"" ( ""Name"" )
-                    VALUES ( 'Unit Testing {Guid.NewGuid()}, Inc.' )
+                _tenantId = await connection.QueryFirstOrDefaultAsync<Guid>($@"
+                    INSERT INTO ""Tenant"" ( ""Name"", ""DefaultAssetTypeId"" )
+                    VALUES ( 'Unit Testing {Guid.NewGuid()}, Inc.', 1 )
                     RETURNING ""Id"";");
 
                 // Get a User to use
