@@ -1,7 +1,7 @@
 ﻿@*
     UNIT TEST Default Invoice Template for Dash Accounting System v2.0
 *@
-@model DashAccountingSystemV2.Models.Invoice
+@model DashAccountingSystemV2.BackEnd.Models.Invoice
 <!DOCTYPE html>
 <html lang="en">
 <head>
