@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Console;
 using Microsoft.Extensions.Options;
-using IdentityServer4.EntityFramework.Options;
 using Moq;
 using Npgsql;
-using DashAccountingSystemV2.Data;
+using DashAccountingSystemV2.BackEnd.Data;
 using DashAccountingSystemV2.Tests.Repositories;
 
 namespace DashAccountingSystemV2.Tests
@@ -59,10 +54,7 @@ namespace DashAccountingSystemV2.Tests
 
             var dbOptions = dbCtxOptionsBuilder.Options;
 
-            var operationalStoreOptions = new OperationalStoreOptions();
-            var wrappedOperationalStoreOptions = new OptionsWrapper<OperationalStoreOptions>(operationalStoreOptions);
-
-            ApplicationDbContext appDbContext = new ApplicationDbContext(dbOptions, wrappedOperationalStoreOptions);
+            ApplicationDbContext appDbContext = new ApplicationDbContext(dbOptions);
             
             await appDbContext.Database.EnsureCreatedAsync();
             await appDbContext.Database.MigrateAsync();
